event-transformer: extract color palette
