+++ src/apps/selection/app.rs
@@ -149,10 +149,8 @@ mod test {
     }
     impl EventTransformer for Transformer {
         fn into_index(&self, _event: Event) -> Result<Option<u16>, Error> { Err(Error::Unsupported) }
-        fn into_color_palette_index(&self, _event: Event) -> Result<Option<u16>, Error> { Err(Error::Unsupported) }
         fn from_image(&self, _image: Image) -> Result<Event, Error> { Err(Error::Unsupported) }
         fn from_index_to_highlight(&self, _index: u16) -> Result<Event, Error> { Err(Error::Unsupported) }
-        fn from_color_palette(&self, _color_palette: Vec<[u8; 3]>) -> Result<Event, Error> { Err(Error::Unsupported) }
     }
 
     const TRANSFORMER: Transformer = Transformer {};
