+++ src/apps/spotify/app/render_state.rs
@@ -167,10 +167,6 @@ mod test {
                 return Err(Error::Unsupported);
             }
 
-            fn into_color_palette_index(&self, _event: Event) -> Result<Option<u16>, Error> {
-                return Err(Error::Unsupported);
-            }
-
             fn from_image(&self, mut image: Image) -> Result<Event, Error> {
                 let mut prefix = Vec::from("IMG".as_bytes());
                 let mut bytes = vec![];
@@ -181,10 +177,6 @@ mod test {
 
             fn from_index_to_highlight(&self, index: u16) -> Result<Event, Error> {
                 return Ok(Event::Midi([index as u8, index as u8, index as u8, index as u8]));
-            }
-
-            fn from_color_palette(&self, _color_palette: Vec<[u8; 3]>) -> Result<Event, Error> {
-                return Err(Error::Unsupported);
             }
         }
 
@@ -228,10 +220,6 @@ mod test {
                 return Err(Error::Unsupported);
             }
 
-            fn into_color_palette_index(&self, _event: Event) -> Result<Option<u16>, Error> {
-                return Err(Error::Unsupported);
-            }
-
             fn from_image(&self, mut image: Image) -> Result<Event, Error> {
                 let mut prefix = Vec::from("IMG".as_bytes());
                 let mut bytes = vec![];
@@ -242,10 +230,6 @@ mod test {
 
             fn from_index_to_highlight(&self, index: u16) -> Result<Event, Error> {
                 return Ok(Event::Midi([index as u8, index as u8, index as u8, index as u8]));
-            }
-
-            fn from_color_palette(&self, _color_palette: Vec<[u8; 3]>) -> Result<Event, Error> {
-                return Err(Error::Unsupported);
             }
         }
 
@@ -292,20 +276,12 @@ mod test {
                 return Err(Error::Unsupported);
             }
 
-            fn into_color_palette_index(&self, _event: Event) -> Result<Option<u16>, Error> {
-                return Err(Error::Unsupported);
-            }
-
             fn from_image(&self, _image: Image) -> Result<Event, Error> {
                 return Err(Error::Unsupported);
             }
 
             fn from_index_to_highlight(&self, index: u16) -> Result<Event, Error> {
                 return Ok(Event::Midi([index as u8, index as u8, index as u8, index as u8]));
-            }
-
-            fn from_color_palette(&self, _color_palette: Vec<[u8; 3]>) -> Result<Event, Error> {
-                return Err(Error::Unsupported);
             }
         }
 
@@ -339,19 +315,11 @@ mod test {
                 return Err(Error::Unsupported);
             }
 
-            fn into_color_palette_index(&self, _event: Event) -> Result<Option<u16>, Error> {
-                return Err(Error::Unsupported);
-            }
-
             fn from_image(&self, _image: Image) -> Result<Event, Error> {
                 return Err(Error::Unsupported);
             }
 
             fn from_index_to_highlight(&self, _index: u16) -> Result<Event, Error> {
-                return Err(Error::Unsupported);
-            }
-
-            fn from_color_palette(&self, _color_palette: Vec<[u8; 3]>) -> Result<Event, Error> {
                 return Err(Error::Unsupported);
             }
         }
