+++ src/midi/device.rs
@@ -5,7 +5,7 @@ use portmidi::{InputPort, OutputPort, MidiEvent, MidiMessage};
 
 pub use crate::image::Image;
 use super::Error;
-use super::features::{AppSelector, GridController};
+use super::features::{AppSelector, ColorPalette, GridController};
 
 #[derive(Clone, Debug, PartialEq)]
 pub enum Event {
@@ -13,14 +13,10 @@ pub enum Event {
     SysEx(Vec<u8>),
 }
 
-pub trait EventTransformer: AppSelector + GridController {
+pub trait EventTransformer: AppSelector + ColorPalette + GridController {
     /// Device that can associate a MIDI event to an unsigned integer,
     /// that can be used to access elements of an indexed collections.
     fn into_index(&self, event: Event) -> Result<Option<u16>, Error>;
-
-    /// Device that can associate a MIDI event to a color from a palette,
-    /// let’s be realistic, this only be used by the paint application.
-    fn into_color_palette_index(&self, event: Event) ->  Result<Option<u16>, Error>;
 
     /// Device that can render an image.
     fn from_image(&self, image: Image) -> Result<Event, Error>;
@@ -28,9 +24,6 @@ pub trait EventTransformer: AppSelector + GridController {
     /// Device that can highlight an element corresponding to an unsigned integer,
     /// this method can be used to emphasize the active element of a collection.
     fn from_index_to_highlight(&self, index: u16) -> Result<Event, Error>;
-
-    /// Device that can highlight buttons that can be pressed to select colors
-    fn from_color_palette(&self, color_palette: Vec<[u8; 3]>) -> Result<Event, Error>;
 }
 
 /// MIDI Device that is able to emit MIDI events
