+++ src/midi/devices/launchpadpro/index.rs
@@ -19,25 +19,6 @@ pub fn into_index(event: Event) ->  Result<Option<u16>, Error> {
     });
 }
 
-pub fn into_color_palette_index(event: Event) ->  Result<Option<u16>, Error> {
-    return Ok(match event {
-        // event must be a "note down" with a strictly positive velocity
-        Event::Midi([176, data1, data2, _]) if data2 > 0 => {
-            // the device provides a 10x10 grid if you count the buttons on the sides
-            let row = data1 / 10;
-            let column  = data1 % 10;
-
-            // we’ll use the bottom row to select colors
-            if row == 0 && column >= 1 && column <= 8 {
-                Some(column - 1).map(|index| index.into())
-            } else {
-                None
-            }
-        },
-        _ => None,
-    });
-}
-
 pub fn from_index_to_highlight(index: u16) -> Result<Event, Error> {
     if index > 63 {
         return Err(Error::OutOfBoundIndexError);
@@ -49,27 +30,6 @@ pub fn from_index_to_highlight(index: u16) -> Result<Event, Error> {
     let led = row * 10 + column;
 
     let bytes = vec![240, 0, 32, 41, 2, 16, 40, led, 45, 247];
-    return Ok(Event::SysEx(bytes));
-}
-
-pub fn from_color_palette(color_palette: Vec<[u8; 3]>) -> Result<Event, Error> {
-    if color_palette.len() > 8 {
-        return Err(Error::OutOfBoundIndexError);
-    }
-
-    let mut bytes = vec![240, 0, 32, 41, 2, 16, 11];
-
-    for index in 0..color_palette.len() {
-        let led = (index + 1) as u8;
-        bytes.append(&mut vec![
-            led,
-            color_palette[index][0] / 4,
-            color_palette[index][1] / 4,
-            color_palette[index][2] / 4,
-        ]);
-    }
-    bytes.push(247);
-
     return Ok(Event::SysEx(bytes));
 }
 
@@ -143,90 +103,5 @@ mod tests {
             .collect::<Vec<Option<u16>>>();
 
         assert_eq!(expected_output, actual_output);
-    }
-
-    #[test]
-    fn into_color_palette_index_given_incorrect_status_should_return_none() {
-        let event = Event::Midi([128, 3, 10, 0]);
-        assert_eq!(None, into_color_palette_index(event).expect("into_color_palette_index should not fail"));
-    }
-
-    #[test]
-    fn into_color_palette_index_given_low_velocity_should_return_none() {
-        let event = Event::Midi([176, 3, 0, 0]);
-        assert_eq!(None, into_color_palette_index(event).expect("into_color_palette_index should not fail"));
-    }
-
-    #[test]
-    fn into_color_palette_index_given_out_of_grid_value_should_return_none() {
-        let events = vec![
-            [176, 00, 10, 0],
-            [176, 09, 10, 0],
-            [176, 10, 10, 0],
-            [176, 11, 10, 0],
-            [176, 12, 10, 0],
-            [176, 13, 10, 0],
-            [176, 14, 10, 0],
-            [176, 15, 10, 0],
-            [176, 16, 10, 0],
-            [176, 17, 10, 0],
-            [176, 18, 10, 0],
-            [176, 19, 10, 0],
-        ];
-
-        for event in events {
-            let event = Event::Midi(event);
-            assert_eq!(None, into_color_palette_index(event).expect("into_color_palette_index should not fail"));
-        }
-    }
-
-    #[test]
-    fn into_color_palette_index_should_correct_value() {
-        let actual_output = vec![1, 2, 3, 4, 5, 6, 7, 8]
-            .iter()
-            .map(|code| into_color_palette_index(Event::Midi([176, *code, 10, 0])).expect("into_color_palette_index should not fail"))
-            .collect::<Vec<Option<u16>>>();
-
-        let expected_output = vec![0, 1, 2, 3, 4, 5, 6, 7]
-            .iter()
-            .map(|index| Some(*index))
-            .collect::<Vec<Option<u16>>>();
-
-        assert_eq!(expected_output, actual_output);
-    }
-// 2. values are divided by 4
-
-    #[test]
-    fn from_color_palette_when_too_many_colors_then_return_out_of_bound_error() {
-        // a color palette of nine items should not be supported (even if they’re all black)
-        let color_palette = vec![[0, 0, 0]; 9];
-        let actual_event = from_color_palette(color_palette);
-        assert_eq!(actual_event, Err(Error::OutOfBoundIndexError));
-    }
-
-    #[test]
-    fn from_color_palette_when_valid_palette_then_divide_all_values_by_four() {
-        let color_palette = vec![
-            [12, 24, 48],
-            [96, 16, 36],
-            [8, 192, 56],
-        ];
-
-        let actual_event = from_color_palette(color_palette).unwrap();
-        assert_eq!(actual_event, Event::SysEx(vec![
-                // Prefix for "bluk lighting" a set of LEDs
-                240, 0, 32, 41, 2, 16, 11,
-                // Identifier for the first LED
-                1,
-                // The Launchpad Pro only accepts 3-byte colors,
-                // where each byte has a value within the [0; 63] range.
-                3, 6, 12,
-                // Identifier and color for the second LED
-                2, 24, 4, 9,
-                // Identifier and color for the third LED
-                3, 2, 48, 14,
-                // Suffix for LaunchpadPro SysEx commands
-                247,
-        ]));
     }
 }
